Move dynmodb lock into its own crate.
