+++ rust/Cargo.toml
@@ -72,13 +72,19 @@ async-trait = "0.1"
 rust-dataframe-ext = []
 datafusion-ext = ["datafusion"]
 azure = ["azure_core", "azure_storage", "azure_identity", "reqwest"]
-s3 = ["rusoto_core/native-tls", "rusoto_credential", "rusoto_s3/native-tls", "rusoto_sts/native-tls", "rusoto_dynamodb/native-tls", "maplit"]
-s3-rustls = ["rusoto_core/rustls", "rusoto_credential", "rusoto_s3/rustls", "rusoto_sts/rustls", "rusoto_dynamodb/rustls", "maplit"]
+s3 = ["rusoto_core/native-tls", "rusoto_credential", "rusoto_s3/native-tls", "rusoto_sts/native-tls", "rusoto_dynamodb/native-tls", "maplit", "dynamodb_lock/native-tls"]
+s3-rustls = ["rusoto_core/rustls", "rusoto_credential", "rusoto_s3/rustls", "rusoto_sts/rustls", "rusoto_dynamodb/rustls", "maplit", "dynamodb_lock/rustls"]
 gcs = ["async-stream", "tame-gcs", "tame-oauth", "reqwest"]
 glue = ["s3", "rusoto_glue"]
 
 [build-dependencies]
 glibc_version = "0"
+
+[dependencies.dynamodb_lock]
+path = "../dynamodb_lock"
+version = "0"
+default-features = false
+optional = true
 
 [dev-dependencies]
 utime = "0.3"
